fix merge conflicts, cyclical dependencies, and various bugs in release candidate
