+++ services/metadata_service/api/metadata.py
@@ -1,6 +1,8 @@
 from aiohttp import web
 import json
-from services.utils import read_body, format_response, handle_exceptions
+from services.utils import read_body
+from services.metadata_service.api.utils import format_response, \
+    handle_exceptions
 import asyncio
 from services.data.postgres_async_db import AsyncPostgresDB
