+++ services/metadata_service/api/run.py
@@ -1,6 +1,8 @@
 import asyncio
 from services.data.models import RunRow
-from services.utils import read_body, format_response, handle_exceptions
+from services.utils import read_body
+from services.metadata_service.api.utils import format_response, \
+    handle_exceptions
 from services.data.postgres_async_db import AsyncPostgresDB
